Added a Privacy Panel as a wikitext code
In some countries (as in Europe) visitors should have a way to opt out; Piwik provides
this feature through an iframe itself setting an exclusion cookie. To embed this iframe
in MediaWiki, a specific parser tag <piwikPrivacyPanel /> is added; it can be used
typically on the Privacy page so that the visitor can opt out analytics from Piwik.